p4i integration Integrate 117782 from 3 to main: Integrate 117781 from 3.2 to 3: fix scale factor for printing. We were using integer division in one place instead of doubles, leading to rather large rounding errors in some cases (100/34...)
Fixes task #29498

Reviewed by Brad and Volker

Manual test in manualtests/fonts/printing/

[git-p4: depot-paths = "//depot/qt/main/": change = 117783]

+++ src/kernel/qfontdatabase_x11.cpp
@@ -1421,7 +1421,7 @@ QFontEngine *loadEngine( QFont::Script script,
 	    && QPaintDeviceMetrics(fp->paintdevice).logicalDpiY() != QPaintDevice::x11AppDpiY()) {
 	    double px;
 	    XftPatternGetDouble(result, XFT_PIXEL_SIZE, 0, &px);
-	    scale = request.pixelSize/px;
+	    scale = (double)request.pixelSize/px;
 	}
 	fe->setScale( scale );
 	return fe;
@@ -1457,7 +1457,7 @@ QFontEngine *loadEngine( QFont::Script script,
     }
     if (fp && fp->paintdevice
 	&& QPaintDeviceMetrics(fp->paintdevice).logicalDpiY() != QPaintDevice::x11AppDpiY())
-	scale = request.pixelSize/px;
+	scale = (double)request.pixelSize/(double)px;
 
     xlfd += QString::number( px ).latin1();
     xlfd += "-";
